Scala 3.5.0 (#196)
* Scala 3.5.0

* chore: update Scala version to 3.5.0

* chore: update laminarVersion to 17.1.0

Vars <3

* feat: Update password field rendering in UI5WidgetFactory

This commit updates the `UI5WidgetFactory` in the `modules/ui5` directory to include a new implementation for rendering a password input field. The `renderSecret` method now uses the `Input` component from the `com.raquo.laminar.api.L` package and sets the input type to "password". This change allows for rendering password fields with the appropriate input type in UI5 widgets.

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* feat: Update rendering of form fields in UI5WidgetFactory

This commit updates the `UI5WidgetFactory` in the `modules/ui5` directory to improve the rendering of form fields. The `renderVar` method now uses the `asForm` extension method from the `Form` object to render the form fields. This change ensures consistency in rendering across different types of form fields in UI5 widgets.

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* feat: Update dependencies and add scala-java-time library

This commit updates the dependencies in the `build.sbt` file. It updates the `web-components-ui5` library to version 2.0.0-RC1 and adds the `scala-java-time` library with version 2.6.0. These updates ensure compatibility with the latest versions of the dependencies and introduce new functionality provided by the `scala-java-time` library.

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* feat: Update dependencies and add scala-java-time library
Date support

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>