feat: Add sealed classes example to client code
The code changes in this commit add a new file `Sealed.scala` to the `examples/client/src/main/scala/samples` directory. This file contains an example of using sealed classes in Scala. It defines a sealed trait `Animal` and two case classes `Horse` and `Lama` that extend the `Animal` trait. It also includes a case class `Owner` that has a `name` and a `pet` field of type `Animal`. The commit also updates the `HelloWorld.scala` file to include the new `sealedClasses` sample component.

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>