feat: Add new sample for opaque type in Index.scala
Add a new sample for opaque type in the Index.scala file. This sample demonstrates the usage of an opaque type in a case class called Person, which includes a name and a password. The sample also includes a simpleVar variable that holds an instance of Person and renders it in the UI.